docs: second pass on upgrade guide
Added short version of the manual upgrade guide, and adjusted the guide
to a simpler process. Now instead of having users migrate to `0001`,
we have them go to `zero` and then `--fake` migrations back up to latest
after upgrades. This is a cleaner process overall and allows for some
flexibility in case someone is following the guide at a later date,
when/if there are actual changes to concrete models in `mentions` app.