Enabled alternative format for settings.
Options may now be applied using a `dict` structure, if preferred.
Warning: Both formats are equivalent but you must use one or the other,
not both. If WEBMENTIONS is defined, it takes precedence and
WEBMENTIONS_ keys will be ignored.

The following are equivalent:

WEBMENTIONS = {
    "TIMEOUT": 10,
    "USE_CELERY": True,
}

WEBMENTIONS_TIMEOUT = 10
WEBMENTIONS_USE_CELERY = True